src/Library/GN.Library.AI/WebScraping/py/WebScraper/requirements.txt
requests 
beautifulsoup4 
chromadb 
lxml
pydantic 
argparse
nats-py
fastapi
uvicorn
sentence-transformers
